fix(chrome): use sub-minute heartbeat when gap is 1
compute_sessions joins only when spacing is strictly less than the gap,
so a 1-minute cadence with --gap-minutes 1 still split sessions. Clamp to
0.5 minutes in that case and cover it with a regression test.

+++ core/runtime_collectors.py
@@ -79,7 +79,7 @@ class RuntimeCollectors:
             profiles, dt_from, dt_to, self.home, self.classify_project, self.make_event
         )
 
-    def _web_visit_collapse_minutes(self) -> int:
+    def _web_visit_collapse_minutes(self) -> float:
         collapse = 12
         gap = 15
         if self.cli_args is not None:
